requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.10.1
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
enum34==1.1.10
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
namedlist==1.8
outcome==1.2.0
packaging==23.1
pipdeptree==2.7.1
pluggy==1.0.0
py==1.11.0
PySocks==1.7.1
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==2.0.4
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
wsproto==1.2.0
